mode of payment
update ui
